mf_modernization/requirements.txt
openai
loguru
python-dotenv
streamlit
google-generativeai
streamlit-option-menu
ollama
fastapi
uvicorn
python-multipart
pyyaml
requests
markdown
markdown-it-py 
xhtml2pdf
markdown2
xhtml2pdf

docx2pdf
pypandoc
python-docx
pdfkit